backend/palindrome/requirements.txt
google-cloud-storage
sqlalchemy
pymysql
